check class representations before splitting and creating labels
I am not completely sure though that is correct to remove the class from the list of the categories whenever there is not a single data point representing it, e.g. if there are other databases containing enzymes of that class which will survive the preprocessing step. Then our algorithm has 0.0% chance inferring that class.