Simplified networks, plotting and training
Now we use sigmoids (instead of the awkward 2-way softmaxes per
class), we plot the per_class_metrics separately and have a plot
with single acc-cy and loss curves for all classes. And also the
model can use the same data directory, but be trained for different
enzyme classes as long as all data is present.